fix: improve accessibility in several areas
I made the change to the JavaScript file to delegate the actual event
happen on the parent element of the image rather than the element itself
to avoid having to change a bunch of stuff with ARIA and HTML.

The hover effect should be self-explanatory. I've also added the same
outline effect for when you focus on the element using a keyboard.